pyproject.toml
[project]
name = "comfyuiapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "uuid>=1.30",
    "uvicorn>=0.35.0",
    "websocket-client>=1.8.0",
]
